IF EVT: Allow intf type specific context to be passed
Allow interface specific context to be passed in as part of the
interface allocation call and thereby allowing customised initialisation
for different types of interfaces. One such example could be for
hardware ports to pass in port id so it can be set on the interface
before generating START event.

VRVDR-46131