npf: set cache-empty flag after an ICMP host redirect
When a packet is sent back out the same interface that it received it on,
an ICMP host redirect is sent out while the original packet is still
being handled. This redirect was processed by NPF, which overwrites
the NPF cache information of the original packet with that of the ICMP
host redirect. After the ICMP host redirect has been sent, it continues
with the original packet with the NPF cache containing the information
on the ICMP host redirect packet, which causes NPF to not handle the
packet correctly.

To rectify this, the code is being updated to set the npf_flag
NPF_FLAG_CACHE_EMPTY if an ICMP redirect is sent and the original packet
is not discarded. This indicates when setting up the NPF cache that it
should be reloaded.

VRVDR-53871