Added range check for time to byte conversion
Queue depths and WRED thresholds are configured in msec
from CLI and converted to bytes in DP. The conversion depends on
the bandwidth of the interface and the calculated bytes can exceed
the range. Added range check against the byte range defined in
vyatta-policy-qos-groupings-v1.

Updated QoS config protocol number

VRVDR-49187
VRVDR-46362
VRVDR-53916