LAG: Distinguish between a member being usable and enabled
A LAG member may be enabled by teamd (control plane for LAG intf)
but then subsequently be marked as unusable by other protocols such
as micro BFD, a fault detection protocol. Therefore need to maintain
state for both.  This information will then be used to decide whether
a member will added to the TX HASH for the LAG.  Also, when a new
member is added, assume it is usable till told otherwise.

VRVDR-46131