fal: don't signal routes with non-main tables to FAL plugins
There is no point in hardware switches in having two types of tables,
so any reasonable hardware will just have one type of table:
VRFs. They may give some flexibility in their use which could emulate
PBR tables, but having this support in the route APIs is an
unnecessary complexity that is unlikely to have any benefits. For this
reason, block additing of routes in non-main tables.

VRVDR-53924