fal: use new FAL route APIs if supported
Use the new FAL route APIs if possible, falling back to the old,
deprecated APIs if not supported for backwards compatibility.

Change the IPv4 and IPv6 route programming to pass through the VRF FAL
object, since we don't want to look this up in the FAL translation
layer for performance reasons.

VRVDR-53924