fal: use new FAL IP neighbor APIs if supported
Use the new FAL IP neighbor APIs if possible, falling back to the old,
deprecated APIs if not supported for backwards compatibility.

Change the IPv4 and IPv6 neighbor programming to pass through the
router interface FAL object, since we don't want to look this up in
the FAL translation layer for performance reasons.

Fixes: VRVDR-53924