micro bfd: allow micro BFD packets when no address
Even if we don't have an IPv6 address, we still need to be able to handle
some layer 4 traffic. One example is micro BFD which exchanges packets
on using a link-local MAC address. Since this is on a LAG member, the
interface, being a LAG member, will never have an IPv6 address and must
pass traffic to allow an IPv6 address to be negotiated for the LAG.

Bugfix: VRVDR-46131