ARP and ND: entry should not be valid after failure notification
Both ARP and ND leave a valid entry in their dataplane caches after
receiving a NUD failure notification from the kernel. The reason is
that there is no LL addr for NUD_FAILED, but there is no longer any
entry update due to the coverity fix VRVDR 42250. The solution is to
revert this fix.

VRVDR-54041