ARP: restrict handling of notifications from kernel
While unlikely, ensure that no INCOMPLETE entries are added to the ARP
cache. Ignore any notifications for STALE, DELAY or PROBE state, as
valid entries are added based on REACHABLE state. Handle NUD fail
notifications as before, but ensure that only existing entries are
updated. This handling is consistent with that for ND.

VRVDR-54041