ARP: remove entry after failure notification
ARP removes the valid flag from an entry after receing a NUD fail
notification from the kernel. The entry reverts to be being incomplete
('show arp' displays it as PENDING), as there is no requirement to keep
track of the failure notified by the kernel. There is no advantage in
keeping this entry and the show output could be misinterpreted. So
delete it, this handling being consistent with that for ND.

VRVDR-54041