GPC: create a FAL policer when necessary
Create a FAL policer object when one is configured.  Save its object-id
so it can be used with a new pmf_rule pp_action.qos_policer action.

Updated the gpc op-mode code to only show the configured fields of a
policer, this in turn required a small change to the dp_test_gpc_pb.c
unit-test.

Updated fal_plugin_policer.c so that the FAL unit-tests can cope with
an ingress policer.

Part of VRVDR-46091