lag: preserve started state across reconfigure
When a LAG interface switches into jumbo mode, it is necessary to
reconfigure the member PMDs so that they also support jumbo mode.
It's important to preseve the started state across this reconfiguration
because later operations preserve the started and will incorrectly see
that the interface is stopped and preserve that state. For some PMD types,
the bonding interface always stops/starts the members and hides this bug.

Bugfix: VRVDR-53928