Update soft_ticks using clock_gettime (re-factored)
VRVDR-54042. clock_gettime is called every 10mS, with parameter
CLOCK_REALTIME_COARSE in order to update an epoch timestamp.  This is
used to provide a more accurate soft_ticks value.   We also provide an
uptime in microsecs (soft_ticks_us) and a unix epoch timestamp in
microsecs, unix_epoch_us.