CRYPTO: Do not set xfrm_direct true
Do not set xfrm_direct true until a message has been received over the
xfrm_direct channel. When the dataplane is restarted and restart is
issued there is a window when xfrm_direct is true, and an xfrm_ack is
generated, before xfrm_direct is cleared by a message being received
over the vplaned channel. In the longer term the only messages
received will be on the xfrm_direct channel.

VRVDR-53932