clang: fix variable length array size warnings
Delay the stack allocation of variable arrays until we verify that we
will not make a zero length allocation. In two cases, it necessary to
convert a range check to use a similar expression that doesn't confuse
the clang-tidy analyzer. While *name and strlen(name) > 1 are the
fundamentally the same check, it seems to confuse the static analyzer.

Bugfix: VRVDR-54147