rcu: fix compiler errors with pointers and integers
The static checking in later compilers complains about the assignment
between pointers and integers and the user-space RCU casting does not
hide this issue sufficiently:

    src/if.c:3299:3: error: comparison between pointer and integer ('fal_object_t' (aka 'unsigned long') and 'void *') [-Werror]
        rcu_assign_pointer(ifp->fal_l3, fal_l3);
        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

For the FAL, object ids are expected to be opaque pointer as far as
the application is concerned. CMM_{LOAD,STORED}_SHARED should be
sufficent here.

Bugfix: VRVDR-54202