ACL: Add new summary bits for actions
Jira: VRVDR-54154

These are for operating a hardware QoS policer, setting the
colour for a hardware QoS queuing mechanism, and setting a
hardware 'designation' which is meaningless as a software
abstraction.

Amongst other issues exposed by this addition include that
the 'designation' bit is very much specific to one hardware
platform.  So these new bits, and the whole use of the
summary bits for actions should be revisited.

At a minimum, the match criteria summary bits should be in
a distinct set from the actions.  It is also necessary to
change how the counter action summary bits are used.