ACL: Use new summary bits for encoding new actions
Jira: VRVDR-54154

Pass the use of the policer action to the creation of the FAL
group/table object, by using the new summary bit.  Change the
creation of the FAL rule/entry object to also use the same
summary bit.

Pass notification of Colour and Designation in the FAL group
creation by use of new summary bits for them.