fw: skip NAT processing with session in wrong direction
If NAT is performed on a session looked up in the wrong direction it
results in it translating the wrong address and also causing an incorrect
checksum. This affects both internal hairpin and tunnel-to-self set-ups.

The code has been updated to skip DNAT and SNAT processing if the session
is in the wrong direction.

This will still work for return traffic, as if there is a DNAT rule
(which is run on 'input') with the return traffic it will look up the
'backward' DNAT sentry in the 'output' direction, and this is not marked
as internal-hairpin as it is the expected direction for that sentry. It
is only marked as internal-hairpin if the direction is not the expected
one (i.e. the DNAT 'forward' sentry matching on output, or the DNAT
'backward' sentry matching on input).

VRVDR-54027