GPC: add rule action code and empty match code
Start to construct a pmf_rule struct for each rule in a feature table.
Initial add the code for the actions, but only add empty functions
for the matches (which will be added in the next commit to get
around the 400 line rule).  These pmf_rule structs are used by pmf_hw.c
to build the FAL objects for each rule.

Part of VRVDR-46091