crypto: discontinue policy-rule tags
The NPF ruleset based required policy-rule tags,
to combine a rule index and a priority, since the
priorities in the IPsec SPD are not unique.

RLDB allows to specify a rule-index and priority
individually.

This requires that the IKE control-plane manages an unique
rule-index for each policy. The rule_index field in the XFRM
policy structs need to be set by the IKE control-plane.

VRVDR-53713