nat: only call map_rcu_free sanity if from GC function
Only perform a sanity check if marked as dead, as it means it was called
from the GC routine. It is not called if released for other reasons,
which happens when forcing releases of the resources even if they are
in use. This currently occurs when the dataplane is shutdown or due to
unit tests.

VRVDR-54238

(cherry picked from commit d2992b039eda510870408c3af6625816ffd570dc)