fw: improve comment at "result:" in npf_hook_track()
It is not clear that the comment after "result:" in npf_hook_track()
relates to the too-big ICMP error, so move it within the "if" to make
this clear, and also remove extra information which is not checked here.

Add a comment just after "result:" to say why it might jump there.

VRVDR-54027