tests: fix path redirection checks
The current check allows matches in both directions. If the plugin
path is /usr/foobar then /usr matches. Add a starts_with that better
represents what we want. This is still slightly wrong since the checks
should include a trailing slash. However, the /run and the plugin path
are unique enough that this isn't likely to ever be a problem.

Bugfix: VRVDR-54471