ACL: Prepare to move op-mode support
Jira: VRVDR-54566

Prepare to move the code for 3 commands in to a new file, as this
is not specific to the string based "attached rule group" logic
associated with the file where these currently reside.

Rename the functions in line with the fact that they are primarily
related to the ACL (IPPF) functionality. These commands are:

  - dump data structures (vplsh only, not modelled)
  - show acl counters
  - clear acl counters

The dump command has been adjusted, and an API added, such that it
can safely dump non ACL structures.  This is achieved by only
extracting ACL specific fields from behind the new API if the
elements in question are ACL instances.