GPC: add initial counter support
Use the new GPC counter APIs to create a counter group for each GPC
group and add GPC counters to the counter group.

Update the GPC op-mode code to retrieve the counter values for
each rule if the rule has a counter associated with it.

Added the "features" JSON list to the op-mode JSON so that it is
compatible with the GPC op-mode yang.

Also had to update the GPC protobuf UT to allow for the changes in
the returned JSON.

Part of VRVDR-54743