nh_common: avoid route/neighbor update races
It is possible that the hash table is going away or has gone away
due to VRF changes, but has not yet fully disappeared. After we take
a reference we need to ensure that it isn't NULL during deletes. If
it is, we no longer have a path or interface for this neighbor update
(that will be cleaned up after the RCU expires).

While here fix an oversight with atomic updates of the next hop table
entry.

Bugfix: VRVDR-53960
(cherry picked from commit 27c10647868fbb181cbab770600424fe4e4f4fc9)