GPC: add support for resettable counters
Add a set of 'shadow' counters that hold the value of counters when
the 'gpc clear' VPLSH command is processed.

The counter values returned by the 'gpc show' VPLSH command is the
difference between the current counter values, and the shadow counters.

Also updated the GPC op-mode protocol version since the dataplane now
supports the 'gpc clear' VPLSH command.

Part of VRVDR-54743