npf: Allow unit-tests to reset session ID to 0
The session table json returned from the dataplane uses the session ID
to name the json object for each session.  If a test compares that json
against expected output then we need to ensure the session ID for that
test starts at the same value regardless of one or multiple tests being
run.
VRVDR-54741