ippf/gpc: add missing refcount decrements when counters released
The counter and counter group release functions were not decrementing
the refcount when called. This resulted in counters not getting tidied
up properly when freed, and causing deleted config not being removed,
and also new config not to be added.

This issue was introduced by recent GPC changes: under vrvdr-54566

VRVDR-54749