CRYPTO: Ensuring correct pr->flags handling during an update
Add a new policy rule flag to indicate a policy is being updated. No
flag being set on a policy rule update is an implicit indication it was
an update. This commit makes it explicit.

After a policy has been processed by the npf_cfg_commit ensure
that the flags are cleared.

VRVDR-54747