cgnat: Split up cgn_map_get function #1
Extract the block of code that assigns a new public address and port block
to a new subscriber.  This new function is only used when a new subscriber
and protocol are seen.  For example, if port-block size was 512 then this
new function would be called for the first session, but not for the
subsequent 511 sessions.  This is just moving of code.  There is no
functional change.
VRVDR-54775