cgnat: Split up cgn_map_get function #2
Extract the block of code that assigns a new port-block from an already
assigned public address.  This new function is only used when the current
port-block of a subscriber has been fully allocated, which will only
occur very infrequently.  This is just moving of code.  There is no
functional change.
VRVDR-54775