GPC: delay FAL programming until interface ready
There is a race condition after boot between the GPC config being
programmed down in the FAL, and the interfaces having their fal_l3
feature mode enabled.  If the GPC config occurs first, the FAL
programming fails as the interface aren't ready.

The underlying gpc_hw code already checks whether or not the interface
is ready to accept L3 features in the FAL, and not programme them if it
isn't.  The missing part was not registering for if-feature-mode-change
events, so that interfaces that were slow in getting ready can be
programmed once they are ready.

When one of these events happens, the GPC code walks its config tree
looking for any GPC tables that are attached to the interface that has
changed its mode, and carries out any necessary FAL programming.

Fixes: VRVDR-54766