fal: add FAL_RET_PLUGIN_CONSUMED framer type
A FAL plugin backplane framer may want to completely handle the incoming
packet itself.  In some scenarios, the incoming packet may need special
treatment and redirection to another part of the network stack.  To avoid
adding addtional logic to the dataplane for this, allow the FAL plugin
to do the necessary work.

Bugfix: VRVDR-53937