DPI: Avoid concurrent access of ndpi_flow_struct by multiple lcores
For each flow, an instance of `ndpi_flow_struct` is maintained in the
respective row in the connection table. The `ndpi_flow_struct` structure
contains `ndpi_packet_struct` (`packet` in `struct ndpi_flow_struct`)
which holds per-packet data and is updated by nDPI library for each
packet of the session handled by it. The per packet data includes
pointers to L3, L4 headers and L4 payload which are contained inside the
`rte_mbuf` of the given packet.

Since the connection table is global and if there are not any explicit
locks, any of the lcores can access it at any time.

Under such circumstances, if the forward and backward packets from the
same flow are handled by different lcores, it is possible that the
`ndpi_flow_struct` of the session is concurrently accessed by the two
lcores.

Fixes DAN-316

Co-developed-by: Anshul Thakur <athakur@cdot.in>
Signed-off-by: Anshul Thakur <athakur@cdot.in>
Signed-off-by: Shubham Shrivastava <shubhams@cdot.in>