fwsync: pack a firewall session using protobuf
Serialize a session to a packed protocol buffer.

A packed buffer format looks like:
+--------------------------------------------------+
| len | version | flags | type | packed session(s) |
+--------------------------------------------------+

This packed sessions are used to restore the session on a different router.

This function provides two different types of packing:

- SESSION_PACK_FULL: pack a session and its associated sessions into a
  buffer.
  dp_session_restore() can fully recreate the packed session later with
  same state and statistics as the original session was at the time of
  packing.

- SESSION_PACK_UPDATE: only pack the statistics and states of a session
  along with the lookup keys. This update buffers can be used update the
  states of a session restored earlier.

This commit only handles the packing of firewall sessions. NAT, ALG
sessions are not supported.

VRVDR-48632