gpc: add request for hardware commit after processing GPC protobuf msg
When GPC configuration was changed, the hardware was not being told to
"commit" the changes (via gpc_hw_commit()), and was resulting in the QoS
config not being applied. This adds a call to commit after processing
the GPC protobuf message.

Note that the hardware commit would occur when config changes were made
under the "interfaces" node, as the npf YANG registered a call to script
"end-npf-interfaces", which sent down a "npf-cfg commit" cstore message,
and that called gpc_hw_commit().

VRVDR-54751