GPC: prevent crash when GPC rules not contiguous
The agreed contract between the GPC VCI/protobuf code was that the
VCI/protobuf code would compress non-contiguous rule numbers, so
CLI rules numbers 1, 3, 6 and 9 would become 1, 2, 3 and 4 in the
protobuf.   The GPC dataplane code allocates a rules-table based
upon the number of rules in the protobuf, assuming that the rule
numbers have been compressed.  When the rule-number are sparse
the GPC dataplane code indexes beyond the end of the allocated
rules-table.

The compression of rule-numbers now takes place within the GPC
dataplane code.

Fixes: VRVDR-54893