Separate check patch and commit size check
There is nothing checkpatch specific about the size check, and it would
be useful from CI Actions to be able to run separately. Specifically
when running checkpatch from a Github Action no wrapper is needed and
default action used, with a separate job defined that just runs the size
checker.

VRVDR-54972