cgnat-alg: Add hooks for ALG data/child packets flows
Packets that do not match a CGNAT session are looked up in the ALG
"pinhole enabling" table.  If a match is found then this marks an
ALG data/child flow, and a session will be created.  The pinhole table
entry will contain a pre-allocated mapping for use by the new session.
VRVDR-55072