ut: add s2s configure functions callable from plugins
Added library functions to configure site-to-site IPSec, including
configuring interfaces, policies, and security associations.
A configuration structure is used to specify the IPv4 or IPv6 addresses
and other configuration parameters, and that is passed into functions
so they can extract configuration values they require. The functions
added to the library are:

    _dp_test_s2s_add_vfp_and_bind
    _dp_test_s2s_del_vfp_and_unbind
    _dp_test_s2s_setup_interfaces
    _dp_test_s2s_setup_interfaces_finish
    _dp_test_s2s_teardown_interfaces
    dp_test_s2s_common_setup
    dp_test_s2s_common_teardown

The macros added are:

    dp_test_s2s_setup_interfaces
    dp_test_s2s_setup_interfaces_finish
    dp_test_s2s_teardown_interfaces
    dp_test_s2s_teardown_interfaces_leave_vrf

VRVDR-39228