crypto: add hot marking and unroll loop
This routine is called repeatedly in the RX path for every PMD, so we
need to ensure that the global and routine stay "hot". Manually unrolling
the very short crypto loop also provides a slight performance gain since
the testing for end of loop is 1/3 of the total branch tests.

Bugfix: VRVDR-54129