ecmp: force IPv6 resolution (ND6) for each ECMP member
On hardware forwarding platforms (QAX/J2) the ECMP selection algorithm
may be different to the algorithm used by the dataplane. Consequently
when the hardware requires neighbour resolution for a particular
output path it punts, the dataplane selects a *different* output path
& issues the NS request. With no resolution for its selected path the
hardware continues to punt, that is, endless software forwarding.

Rather than attempt to emulate the hardware selection algorithm, try
to resolve all members of the set.

Fixes: VRVDR-53410