clang: quiet warning about uninitialized value
clang's static analyzer appears to get confused when handling the
block copying that is implemented in rte_ring:

    /usr/include/dpdk/rte_ring_elem.h:295:3: note: Loop
                     condition is false. Execution continues on line 301
        for (i = 0; i < (n & ~0x3); i += 4, idx += 4) {
        ^
    /usr/include/dpdk/rte_ring_elem.h:301:3: note: 'Default' branch
                                  taken. Execution continues on line 301
        switch (n & 0x3) {

clang has assumed that n is some positive non-negative value, but that n
also does not match either of the conditions above. If that did happen,
rte_ring_sc_dequeue_burst could return a positive count and fail to
return ring entries.

To avoid the static warning, disable the warning with NOLINTNEXTLINE.

Bugfix: VRVDR-55125