cgnat-alg: Add function to transfer a mapping
A mapping structure is used to hold an address and port reservation
until such time as it can be used to create a session.  This is a small
structure that is copied between other enclosing data structures
(e.g. pinholes) in order to avoid the complication of mallocs, frees,
pointers and ref counts.
VRVDR-53803