npf-rte-acl: Move trie context creation to separate function
A trie context can be created either from a memory pool or from
the heap depending on the number of rules. The memory pool method
is used during startup to set up a pool of tries to use during
tunnel setup. The heap allocation method is used during a trie
merge.

VRVDR-53712

Co-developed-by: Daniel Gollub <dgollub@att.com>
Signed-off-by: Daniel Gollub <dgollub@att.com>