Dataplane crash in fal_tx_pkt_burst
This crash occurs in the unlikely case of an interface being brought
up before the backplane interface it is bound to comes up, and if there
is switch port tx traffic queued to the backplane interface in this
brief interval. A sufficient solution is to check that the entry for
the backplane interface has been added to ifport_table, as the packets
will be requeued later.

VRVDR-55142