ut: re-enable crypto_perf_scale for full runs
Reduce the number of policies to have an acceptable runtime on
slow test hosts. As well as in preparation to not exceed the ACL
batch limit with the upcoming sub-trie merge logic.

Each policy installation results in an individual commit, which is
a worst-case scenario with regards in piling up merge-tries within
one ACL optimization thread interval.

VRVDR-53712