npf-rte-acl: implement pending_deletes
While a merge of frozen tries is in progress on the acl-opt thread,
delete requests of rules inside that new merged trie need to be removed.

Collect all incoming deletes and put them on the pending_delete list of
this match_ctx. The acl-opt will take those pending_deletes in account
after the initial build of the merged trie.

This operation is protected via the merge lock.

VRVDR-53712