VM interface speed and duplex incorrectly set on KVM
Upgrading from DPDK 19.11 to DPDK 20.11 has resulted in speed/duplex
for virtual interfaces changing from 10g/full to unknown/half. While
versions of qemu supporting VIRTIO_NET_F_SPEED_DUPLEX allow these to be
set, it is not practical to mandate such an upgrade. A solution is to
override the new settings in the dataplane so as to ensure backwards
compatibility.

VRVDR-55194