Fix issue in DSCP remark map on a vif sub-interface
When the egress remark map on the parent interface is removed, the
child interface egress map was also reset to the default egress map
even though it had a specific egress map configured. The issue was
that the parent egress map (default map) was overwritten to the
child interface without checking if it had a egress map configured
for itself.

VRVDR-54822