twamp: process twamp offload control messages
Use a protobuf handler to process the various control messages issued
by the TWAMP daemon in order to establish the UDP-based test session.

A create message is used to establish the flow (addressing, mode,
frame sizes). A counter message is used to request the number of test
packets received & reflected back to the client. This is used by the
daemon to ensure that the dataplane is making progress. Finally a
delete message is used to destroy the flow.

Fixes: VRVDR-53580