QoS: stop requesting unnecessary traffic-class
The design of the QoS hardware code is that it should only create FAL
objects for items that are going to be used.  In the reported problem
the QoS hardware code was trying to create a QoS traffic-class object
that had no QOS queue objects assigned to it.

The FAL code rejected the request to create sched-group for the
traffic-class when it sanity checked the object's attributes.

The spurious request was due to the misplacement of the initialisation
of the local_priority_wrr variable.  Originally this variable was
initialised outside for "for all traffic-classes ..." loop, which
meant that if traffic-class 0 was assigned the local-priority queue,
the local_priority_wrr variable would still be set when it came to
process traffic-class 1.  If traffic-class 1 had no queues assigned to
it, the setting of the local_priority_wrr variable would attempt to
create an unnecessary traffic-class in the FAL.

The fix is to move the initialisation of the local_priority_wrr
variable inside the "for all traffic-classes ..." loop.

Fixes: VRVDR-55318