pipeline: move down ipv[46]-validate enums to allow early features
The feature enums used for ipv[46]-validate start at 1. This means that
a plugin feature cannot request to go before them - doing this fails
with error "no available id for feature".

This updates them to start at 3, which would allow up to 2 features
to be added before. These should be features such as monitoring ones,
which do not change the packet. Space is only left for 2 early features,
as the feature node is currently using 16 bits to track the features,
and so this still leaves space for 3 (IPv4) and 4 (IPv6) late features
after the early enums, and before the no-address/no-forwarding features
which are at the end.

VRVDR-55443