twamp: correct the order of timestamps
The two key timestamps (arrival and departure) were written to the
response packet in the wrong order leading to a negative value for the
"reflector processing time".

Correct the order (departure followed by arrival) and the need for the
error estimate (departure timestamp only).

In order to provide a slightly more accurate processing time, delay
collection of the departure timestamp until the last possible moment.

Fixes: VRVDR-55567