npf-rte-acl: fix spurious lookup rte-acl rule failures
Since a custom hash function is used, which just uses a subset of
rte_acl_rule, we need also a custom compare function.

rte_acl is using a pointer of struct rte_acl_rule, which gets copied
by rte_hash as hash key. This requires that at least rte_acl_rule_data
gets copied to include atlesat the userdata/rule_no field for the
compare function. For that reason the key_len needs to be increased
to the size of struct rte_acl_rule_data.

VRVDR-55605