add inline functions to perform quick hashes
Profiling using "perf record" showed a large amount of type was taken
performing hashing in looking up the sentry table. That was because it
used rte_jhash_*() functions which perform a lot of processing.

Therefore some replacement functions which create a hash by performing
XORing the values are added here.

The functions added are dp_xor_1word(), dp_xor_[234]words(), and
dp_xor_array32(). These are to be used as replacements for the
rte_jhash*() functions for hash tables when being calculated on the
packet forwarding path.

VRVDR-55148