npf: ensure timeout-get functions cope with no dataplane session
Logging using npf_session_tcp_log() or npf_session_gen_log() would cause a
crash if the dataplane session has not yet been created, which can occur
on the log of a session which is being activated due to the s_session
field (struct session *) of an npf_session_t being NULL . There are
checks for this in npf_session_log(), but prior to that timeout values
were looked up and accessed the s_session field.

The functions npf_session_gen_log() and npf_session_tcp_log() have been
updated to cope with s_session being NULL, by passing in the s_session
as a parameter which it now checks, instead of passing in the custom
timeout value.

VRVDR-55721