cgnat: Invalidate npf cache if a CGNAT translation occurs
If npf firewall exists on input and output of a packet flow, then the pkt
will see 1. firewall, 2. cgnat, 3. firewall.  The pkt is npf-cached the
first time in the firewall only. If a CGNAT translation occurs then the
npf-cache must be invalidated before the second time thru the firewall.
VRVDR-55653