add function which notifies features then frees an mbuf
Add function dp_pktmbuf_notify_and_free() which should be called when
a buffer is freed, so that prior to freeing it notifies features which
have registered for the drop of a packet. This allows features such as
NetFlow, which collects information on packets, to know that processing
has completed on the packet and it should end collecting information
and send what it has collected to the exporter.

VRVDR-55339