replace calls to rte_pktmbuf_free() to calls to fn which notifies
Update calls to rte_pktmbuf_free() to call dp_pktmbuf_notify_and_free()
instead, so that features that have registered a function to be called
on packet drop get their functions called. This allows features such
as NetFlow, which collects information on packets, to know that
processing has completed on the packet and it should end collecting
information and send what it has collected to the exporter.

VRVDR-55339