Jumbo MTU capability omitted
When rte_eth_dev_configure() is called, if the MTU on the
interface is > 1500 and the DEV_RX_OFFLOAD_JUMBO_FRAME flag
is not set, the MTU is reset to 1500. The issue can occur
when setting the interface speed when the MTU has already
been set > 1500. Fix is to ensure that the flag is set when
in the capability mask.

VRVDR-55749

(cherry picked from commit 8de4bc74b7e82e85e6cdee3a490f079f02fbf025)