pktmbuf: use nearest NUMA pool for control packets
If port 0 is not enabled, then the associated mbuf pool associated
not be present. This situation is common in Azure environments where
the accelerated interface is owned and used by the virtual interface.
Since we don't configure the accelerated interface, the portid -> pool
association is never made. Add a new routine to get the nearest NUMA
pool and use that pool instead.

Note that one can also choose not to enable ports via a command line
option. This had the potential to trigger the same bug.

Bugfix: VRVDR-55966