Fix issue of missing ndpi session info for NetFlow
The ndpi code is not providing application information for NetFlow when
showing sessions or giving session logs. This is because when NetFlow
is enabled on an interface it does not call dpi_refcount_inc() on the
npdi engine, and the ndpi engine skips returning JSON and logging as it
thinks ndpi is not being used ("ndpi_refcount" being 0).

This is fixed by calling dpi_refcount_inc() on enabling of DPI on
an interface, and calliing dpi_refcount_dec() on disabling DPI on an
interface. It passes in parameter dpi_global_engine() as the DPI engine,
as that matches the one used when dp_dpi_get_app_id() is called.

VRVDR-56028