Generating icmp error message for large ipv6 pkts
Currently, only IPV4 DF is handled for MPLS. MPLS L3 payload
type value is not set according to the payload it carries.
For processing labeled IPV6 datagram which are too big, the
LSR's strip off the label stack entries to obtain the IP
datagram and create an ICMPv6 Packet Too Big Message
and transmit to the source, Reference RFC3032. No
fragmentation is done on the fly during transit across
the network, so LSR's either forwards or discard it.
After the change, if ipv6 is too big, will just forward
error message to the source.

VRVDR-55895