Avoid delay in changing session timeout after a state change
There is a delay in changing the timeout of sessions when their state
changes, as it is only updated when walking the session table. This can
be confusing when looking at session tables, as it shows an incorrect
timeout for the given state. This fix causes the timeout to be changed
immediately on state change, by updating the stored time at the point
of changing the state.

VRVDR-56136