Sync REACHABLE neighbor state to kernel
The dataplane may have better information about reachability
to a given neighbour than the kernel. If so, use it to update
the kernel when the kernel entry goes STALE.
This is particularly useful under conditions of heavy traffic load
when e.g. a dataplane BFD session is up and functioning but
the kernel sends an ARP request whose replies are lost causing it to
enter the FAILED state.

VRVDR-56158