QOS: fix memory leak due to double config
If the QoS config code in the dataplane receives two sets of config
commands for the same interface without an intervening "qos <if-name>
disable", which should never really happen, it attempts to solider on
by deleting the existing qinfo structure, and creating a new one.

Unfortunately this code-path in the hardware side of the house was
missing a few salient steps.  It didn't free up the existing structures
and just overwrote the pointer to the old root structure with the
address of a new root structure, resulting in a memory leak.  Later on
when trying to create the new sw-objects in the QoS object database,
it would find the old structures and give up trying to create the new
ones.

The fix is to correct tidy up existing structures when this unusual
situation occurs by calling the appropriate qos_<codepath>_disable
function.

There was also an occasion when the FAL's object id for the port was
not cleared when the interface goes down.  This could give the wrong
impression that the port was up.

Also took the opportunity to make the qos_dpdk_port and qos_hw_port
functions look more similar by removing unused code and a bit of
realigment.

VRVDR-56203

(cherry picked from commit a0dbd2fc64618e50ec26f0453c97076291bfe4b9)