QoS: remove qinfo from SLIST when disabling policy
When the QoS DPDK and QoS HW modules need to disable a policy that is
attached to an interface, they cannot do it just using their own disable
functions as doing so will leave the qinfo on the qos_qinfos SLIST,
which will become corrupted when the qinfo is freed.

The fix is for the QoS DPDK and QoS HW modules to call the new platform
independent qos_sched_disable that removes the qinfo from the qos_infos
SLIST before calling the platform dependent qos_<platform_type>_disable
function.

VRVDR-56254