Add qos-vlan and qos-if rows to op mode cmd `show platform dataplane`
qos_interface_hw_stats_get and qos_vlan_hw_stats_get are the top level
functions.
pd_show.c/pd_show_cmd_table has a reference to them. They are called when
pd_show.c/pd_show_dataplane iterates over pd_show_cmd_table.

pd_show_cmd_table expects the functions to return a pointer to an array
of uint32_t's.
This array is shown to the user. Where array[0] is full, array[1] is
partial, array[2] is no-res, etc.
The array is defined as a static global variable because it can not be
defined in the function.

VRVDR-55783