sfp: fix issue of incorrect externally calibrated temp
When there is external calibration of temperature, there is a multiplier
(the slope) and an offset (added on) using calibration constants to the
returned temperature.

The code did a division by 256 to get the temperature in units of 1/256
prior to this adjustment with the calibration constants, resulting in
an incorrect temperature.

The fix is to do the division by 256 after the adjustment due to
calibration.

VRVDR-56385