sfp: correct define/enum used to access RX powers array
The enum SFP_CALIB_CONST_MAX was used to access the array of
RX powers, but it is the number of elements in an unrelated enum
sfp_calib_const_type. Instead "SFP_CALIB_CONST_RX_PWR_CNT - 1" should
be used. As it happens, SFP_CALIB_CONST_MAX is 4, which is the same
value as "SFP_CALIB_CONST_RX_PWR_CNT - 1", so this fix does not change
the behaviour.

VRVDR-56385